chore: componentiza modal comentario
